+++ src/js/ui/DOM/Library/object/input/switchLabel.js
@@ -6,7 +6,7 @@ import DOM from "../../../Classes/dom"
 export default class SwitchLabel {
     constructor([a, b, c], label) {
         b = b || (() => { })
-        c = { include: true, ...c }
+        c = { ...c, include: true }
         const s = new Switch(a, b, c)
         const msc = new CardContent(
             new TwoSidesWrapper(label, s[0]),