+++ src/js/ui/DOM/Library/settings/SettingsActLink.js
@@ -5,12 +5,12 @@ import TwoSidesWrapper from "../object/twoSidesWrapper"
 import Icon from "../object/icon"
 
 export default class SettingsActLink {
-    constructor([act, sign]) {
+    constructor([act, sign, custom = false]) {
         if (typeof act === "string") act = [act]
         sign = sign || (typeof act === "string" ? ucFirst(act[0]) : "(...)")
         return new CardList([
             {
-                content: new TwoSidesWrapper(sign, new Icon("chevron_right")),
+                content: new TwoSidesWrapper(sign, new Icon((custom || "chevron_right"), { marginLeft: "15px" })),
                 handler: (typeof act === "function"
                     ? act
                     : () => { Navigation.hash = { module: "settings", params: act } }),