+++ src/js/ui/DOM/Library/style/index.js
@@ -1,9 +1,9 @@
-import Algin from "./algin"
+import Align from "./align"
 import Padding from "./padding"
 import Width from "./width"
 
 export {
-    Algin,
+    Align,
     Padding,
     Width,
 }