+++ src/res/language/en/strings.js
@@ -68,13 +68,9 @@ export default {
             title: "update settings",
             description: `choose preferred behaviour for "New update" notification, so it can distract you less if you wish.
             Press the pictures to see how it acts`,
-            first_time_explanation_1: `Temply is a web-based
-            application. After first time you run it the app saves itself on your device so you don't have to
-            spend your data by re-downloading it everytime you want to use Temply. In that time we are checking for new
-            updates automatically and downloading them if we found such.`,
-            first_time_explanation_2: `since it's the app runs in your browser we can do this only
-            while Temply is on your screen. When the update is ready it waits to be applied right after the app restart.
-            We can inform you about this updates so you will know about changes and enhancements at first hand`,
+            first_time_explanation: "Temply works thanks to a technology called Progressive Web Apps.",
+            learn_more_about_pwa: "Learn more about the impact of PWAs to the update process",
+            pwa_support_link: "https://temply.procsec.top/help/article/en/what-is-pwa#pwa-role-in-update",
             change_notify_way: "show update settings",
             notify_later: "notify only after installing",
             ready: "new update is ready",